Database test implementation.
 Database test implementation.

+++ lib/listview/customer_list.dart
@@ -28,7 +28,7 @@ class CustomerList extends StatelessWidget {
           },
           child: CustomerListItem(
             name: customer.name,
-            rating: customer.rating,
+            rating: customer.rating.toDouble(),
           ),
         );
       },
